+++ src/application.cpp
@@ -24,7 +24,7 @@ void Application::initialize ()
     _initialized = true;
 }
 
-Application::Application (unsigned int width, unsigned int height)
+Application::Application (size_t width, size_t height)
     : _initialized (false), _window (NULL), _running (false), _paused (false),
       _seconds (0), _allFrames (0), _min (0), _avg (0), _max (0)
 {
